requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.52.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeltranslation==0.18.8
django-mptt==0.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.14
whitenoise==6.4.0
